Run the machine code discovery
This change plumbs through all of the binary parsing code and the setup for the
code discovery. It sets up the necessary logging functionality, as well as the
logic for attempting to parse a bytestring into either an ELF or PE.  Currently
only x86_64/Linux is supported.

The resulting discovery state is not yet used for anything.