Implement a Weird Machine Monitor execution feature
This execution feature takes a list of possible Weird Machine addresses and invokes a callback when one is encountered. The idea is that a separate symbolic execution of the Weird Machine can be started once we see that execution would jump to it, enabling us to verify properties of the Weird Machine behavior.

This feature works by inspecting the register state at each call/return (to address call-oriented-programming and return-oriented-programming) and seeing if the instruction pointer is in the provided list of Weird Machine entry points. Note that this check is non-trivial and can require invoking an SMT solver, as the IP value could be constructed via complex operations.