Resolve concrete symbolic syscall numbers properly
Because `ambient-verifier` uses the SMT array memory model, syscall
number values cannot directly be resolved as being obviously concrete—rather,
they look like reads from an SMT array concatenated together. As a result, we
have to employ other means to ask the SMT solver if a syscall number is
symbolic or concrete. To accomplish this, we employ a trick where we:

1. Ask the online solver for a model of the system call number
2. Send a second query that forbids that model and asks for another

If step (2) fails (i.e., is `Unsat`), then we have a concrete syscall number.
Note that even after this patch, we do not handle truly symbolic syscall
numbers—we continue to fail as before.

Fixes #17.