Support kernel-provided user helper functions, override __kuser_get_tls
In AArch32, TLS is implemented by way of a special `__kuser_get_tls` function
that is available in user space by way of a fixed address in kernel memory. See
https://github.com/torvalds/linux/blob/5bfc75d92efd494db37f5c4c173d3639d4772966/Documentation/arm/kernel_user_helpers.rst#kuser_get_tls.
Such addresses are not part of the loaded binary, but they are accessible. To
make `ambient-verifier` aware of such addresses, this patch:

1. Adds a new `functionKernelAddrMapping` to `FunctionABI` which maps special
   kernel function addresses to custom overrides. In the case of AArch32, we
   instantiate `functionKernelAddrMapping` with an override for
   `__kuser_get_tls`.
2. Implements an override for `__kuser_get_tls` that returns a symbolic value
   from a global pointer dedicated to storing the TLS value. This is very much
   in the spirit of `fsbase` and `gsbase`, which serve similar roles in
   x86_64's TLS implementation.
3. Changes `lookupFunction` to consult `functionKernelAddrMapping` before
   attempting to resolve a function address.

See also `Note [AArch32 and TLS]` in `Ambient.FunctionOverride.AArch32.Linux`.

The rubber hits the road in the `aarch32/read.exe` test case, which makes use of
this newfangled TLS technology.

This also bumps the `macaw` submodule to incorporate
https://github.com/GaloisInc/macaw/pull/249. The ability to concretize certain
applications of `URem`/`SRem` ends up being vital to making `aarch32/read.exe`
pass through the verifier without crashing.

Fixes #43.