Convert some baked-in libgd overrides to crucible-syntax
Now that our extensions to `crucible-syntax` are capable of expressing some of
these overrides, let's rip them out of the verifier itself and put them in a
separate `overrides/libgd` directory. For now, I've left the hacky `malloc` and
`calloc` overrides alone, as they will require more work to fully convert to
`crucible-syntax`.

Addresses three-fifths of #40.