Concretize pointers in read/writes when possible
When executing a `MacawReadMem`/`MacawWriteMem` statement (or one of their
`Cond` variants), we now attempt to concretize the pointer read from (or
written to) before proceeding. Often times, these pointers are (1) symbolic,
but easily resolved to be concrete, or (2) symbolic, but with a small range
of possible values allowed. In the case of (2), we can use `what4`'s new
`unsafeSetAbstractValue` function to update the bounds of the pointer offset,
which may unlock further What4 simplifications. In some cases, this can even
avoid the need to consult the SMT solver at all! As a consequence, various test
cases now generate fewer successful goals (since there is no need to discharge
the goals in the first place).

This has the very desirable consequence of generating smaller SMT queries,
which generally results in better performance. I have tested this by running
the verifier on a large-ish `libgd`-oriented program, which previously would
not even terminate without this patch. Even on the small-ish microbenchmarks
checked into the test suite, there are perf improvements:

Before:

```
  tests/binaries/x86_64/function-call-pie.exe:           OK (9.19s)
  tests/binaries/x86_64/function-call.exe:               OK (8.52s)
```

After:

```
  tests/binaries/x86_64/function-call-pie.exe:           OK (4.43s)
  tests/binaries/x86_64/function-call.exe:               OK (3.89s)
```

Not bad!

This spun out of investigation to make the memory model faster—see #12.
While this isn't a complete overhaul of the memory model (as was originally
envisioned in #12), it does manage to make things go faster.