Log source location upon proving goals
Previously, proven goals would not log anything besides the fact that they
were proven, which made it difficult to determine their provenance. After this
patch, we now have:

```
Proved a goal in 0.052096282s seconds
  tests/binaries/aarch32/global-read.exe:0x10170: in the function 'main'
```

Which is much more descriptive.

While I was in town, I also applied the same treatment to goal timeouts and
goals that error out.

Fixes #59.