Consolidate SymBV–concretization machinery in A.Verifier.Concretize
Previously, the `resolveConcreteStackVal` function and the
`Ambient.Verifier.Concretization` had two separate implementations of machinery
to concretize symbolic bitvectors. This patch consolidates all of this
machinery into `Ambient.Verifier.Concretize`. There was a slight impedance
mismatch to resolve, as `resolveConcreteStackVal` throws an error upon failing
to concretize a value, while most of the functions in `A.Verifier.Concretize`
would just return the original argument unchanged if that occurred. Luckily,
this mismatch was fairly easy to resolve by generalizing
`resolveSingletonSymbolicAs` to return an `Either`.

Fixes #64.