A.Syscall.X86_64.Linux: Speed up pattern-match coverage checking
Unfortunately, GHC's pattern-match coverage checker seems to take forever on
the `Ambient.Syscall.X86_64.Linux` module on GHC 8.10 or later, particularly
due to the module's use of many nested `LLVMPointerRepr`/`testEquality`
patterns. On the other hand, `Ambient.Syscall.AArch32.Linux` doesn't suffer
from this pattern. Some investigation reveals that this is because it bundles
up all of the fancy patterns into a single `syscallABIRepr` value, which is
only tested for equality a single time. This trick is quite effective, as on
my machine, this speeds up the time it takes for GHC's pattern-match coverage
checker to complete from several minutes to several seconds.

This patch copies over the trick from the `AArch32` module to the `Linux`
module.