Add overrides for memcpy and memset
These are simple overrides that simply invoke the corresponding functionality
in `crucible-llvm`. Their simplicity is a virtue, as the `glibc`
implementations of these functions are quite complicated and generate
intractable path conditions.

This required a bit of incidental refactoring:

* I now need a sensible `MemOptions` for use in
  `Ambient.FunctionOverride.Overrides`, so I decided to co-opt
  `Ambient.Syscall.Override.syscallMemOptions` for this purpose. Since its
  purpose is now a bit more general, I moved this definition to
  `Ambient.Override` and renamed it to `overrideMemOptions`.
* I needed to add a `ptrToBv8` function alongside `ptrToBv32` in
  `Ambient.Override`, since `crucible-llvm`'s `doMemset` function takes an
  8-bit bitvector as an argument.

The `memset` override in particular will be useful in a subsequent commit in
which I add a test case (related to #68) which needs to set all bytes in a
buffer to zero before receiving a message from a socket connection.

Fixes #55.