Add --entry-point option
This allows changing the verifier's default entry point of `main` to a custom
function name. In the short-term, this will be useful for binaries that lack a
`main` function. In the long-term, we will want the verifier to start at the
ELF entry point function (e.g., `_start`) instead, so having `--entry-point`
will be a convenient fallback in case a particular binary's `_start` function
proves especially troublesome for the verifier.

Fixes #81.