Test suite: Simplify ExpectedGoals field copying
To produce the actual result of a test case run, we previously took the
result and copied over all of the constant values from the original
`ExpectedGoals`. This is somewhat unfortunate, however, as you have to
remember to update this piece of code every time you add a new field to
`ExpectedGoals` (which is often). Instead of doing that, we now take a
simpler route where we take the original `ExpectedGoals` result and
copy over the results of the test case run that could have conceivably
changed, i.e., the `successful`/`failed` fields. With the way the code
is written now, I expect that this will need to be updated far less
frequently.