x86_64: Model FSBASE more faithfully
Previously, we just wrote symbolic bytes into `FSBASE`. This causes
problems for functions that need to set `errno`, however, as the relevant
assembly needs to dereference the pointer in `FSBASE`. To make this possible,
we simply store the value of the `FSBASE` pointer inside of itself. As luck
would have it, this works for both `musl` and `glibc` alike. I've added a
`Note [x86_64 and TLS]` to explain all this.

Fixes #95.