Resolve addresses offsets using the appropriate memory
Previously, the verifier would always resolve address offsets in
`lookupFunction` by using the main binary's `Memory`. This is the wrong thing
to do in the presence of shared libraries, and the verifier only got away with
it previously due to a lucky fluke.

The verifier now does the correct thing by mapping each address offset to the
binary in which it is defined, and then using that binary's `Memory` for
address resolution purposes. For the full details, see
`Note [Address offsets for shared libraries]` in `Ambient.Load.LoadOptions`.
There is a fair bit of refactoring needed to support this change:

* `InitialMemory` now contains an `imBinaryMems` field to store the `Memory`
  values for the main binary and each shared library.
* The `NamedFunction` data type in `A.Verifier` now contains an `nhAddrOff`
  field to store the address of the function. As a result, we can now use
  `NamedFunction` as a common abstraction for constructing both the
  `fcAddressToHandle` and `fcSoSymbolToFnHandle` values.
* Because we now perform address resolution on a binary-by-binary basis in
  `lookupFunction`, it no longer makes sense to use the same `fcRegion` index
  for each lookup. As a result, I've simply deleted `fcRegion`.

Fixes #96.