Replace filemanip dependency with Glob
`filemanip`'s treatment of `**` patterns in the `namesMatching` function
appears to be broken, which blocks an upcoming refactoring. The `Glob`
library's provides the `glob` function, which has the same type as
`namesMatching` but without this bug, so let's just replace `filemanip` with
`Glob`. This is probably better long-term anyways, since the `filemanip`
library is no longer maintained.