Move .so-related tests to their own subdirectories
As noted in #97, test cases that rely on `.so` files are currently placing
_all_ of their `.so` files in the same directory. In the best case, this makes
the test cases take longer, since every such test case will load all `.so`
files in the directory, including ones that aren't needed. In the worst case,
however, this can cause incorrect runtime behavior.

To avoid this, this patch moves all test cases relying on `.so` files to their
own subdirectories, thereby fixing the first bullet point of #97. In order to
make this happen, I had to change the wildcard patterns in the test suite from
`tests/binaries/*/*.expected` to `tests/binaries/**/*.expected` so that it
would match arbitrary levels of subdirectories.

I have also moved `stripped-so.exe.expected` to
`stripped-so.exe.expected-failing` for the reasons described in the second
bullet point of #97.