Lazy symbolic memory initialization
This is a minor overhaul of the `macaw-symbolic` memory model to make
initializing global memory lazy. Previously, the memory model would create an
SMT array to back global memory and assert equalities on every byte upfront,
before simulation ever begins. This is an expensive process that scales poorly
for large binaries, and I have observed this process eating up all of my RAM
for some particularly large examples that I have tried the verifier on.

Fortunately, there is usually a better way. Most of the time, you don't
actually _need_ to add assertions for every byte in a binary, as invoking the
binary will only use a fraction of the global memory. As a result, we can avoid
adding assertions upfront and instead defer this to simulation time, only
adding assertions for particular parts of the global memory on demand. Not only
does this save on time and memory, but it also typically results in quicker
feedback from the verifier in the event that simulation fails.

For the full details of how this works, see `Note [Lazy memory initialization]`
in the new `Ambient.Extensions.Memory` module. Some highlights:

1. We now pass around a revamped version of `MemPtrTable` where the
  `IntervalMap` not only records whether the memory in a particular address
  range is mutable, but also the symbolic bytes that the address range maps to.
  This `IntervalMap` is consulted during lazy memory initialization to figure
  out which bytes to use to populate global memory.

2. `AmbientSimulatorState` now has an additional `populatedMemChunks` field to
   track which address ranges in global memory have already been populated.

3. When performing a concrete read on immutable memory, there's no need to
   populate global memory at all, as we can simply look up the corresponding
   symbolic bytes in the `IntervalMap` in the `MemPtrTable`. This can save
   quite a bit of time, although it requires a fair bit of machinery to make
   it work.

   A knock-on effect of this optimization is that reading from global memory
   now generates fewer goals, since we do not need to defer to the SMT solver
   as much. As a result, the test suite now has fewer goals to solve.

Fixes #102.