Establish a lookup scope for dynamic symbols
Previously, the verifier assumed that all global dynamic symbols across a
binary and its set of shared libraries had unique names. It turns out that this
assumption isn't safe, as it's quite possible for multiple binaries to define
the same global, dynamic symbols. This patch makes the verifier cope with such
situations by establishing a global lookup scope to determine which dynamic
symbol to look up in the event that there are duplicates among different
binaries:

* Instead of simply loading all shared libraries in the `--shared-objects`
  directory, we now carefully load shared objects in the order determined by
  the binaries' `DT_NEEDED` entries. See
  `Ambient.Loader.ELF.DynamicLoader.loadDynamicDependencies`. (In order to make
  this work as expected, I needed to incorporate an `elf-edit` bugfix in
  https://github.com/GaloisInc/elf-edit/issues/30 by bumping the `elf-edit`
  submodule.)
* `elfDynamicSymbolMap` no longer rejects duplicate symbols among different
  binaries, but instead favors symbols from binaries that appear earlier in
  the global lookup scope. For more on what this means, see
  `Note [Dynamic lookup scope]` in `Ambient.Loader.ELF.DynamicLoader`.

Fixes #107.