Remove awful takeSymbolVersionFileName hackery
I originally put this hack into place because `symbol-versions-main.exe` had
dynamic symbols that included leading paths
(e.g., `symbol-versions/libsymbol-versions-bar.so`). But the fact that leading
paths appear in the first place is because I was silly and compiled
`symbol-versions-main.exe` with the wrong order of linker flags. Oops. In fact,
`symbol-versions-main.exe` is broken due the leading paths appearing in the
`DT_NEEDED` entries. You can't even invoke the binary because of this, even
with `LD_LIBRARY_PATH=.` set!

Once I compiled `symbol-versions-main.exe` with the proper flags, the leading
paths no longer appear, thereby obviating the need for
`takeSymbolVersionFileName` and friends. As such, I've deleted them. This makes
the code nicer now and makes a subsequent commit easier to implement in the
future.