Make `open` syscall override use verifier's `loadString`
Now that the verifier has a version of `loadString` that is aware of the
lazily initialized memory model, fixing #126 is as simple as using that version
of `loadString` in the syscall override for `open`. Most of this patch simply
plumbs arguments through to various places to make that happen.

Fixes #126.