Allow .cbl overrides to invoke functions within the same file
Previously, we weren't registering any functions defined in a `<name>.cbl` file
besides the `@<name>` file itself, which meant that trying to invoke other
functions from `@<name>` would result in an error. We now accomplish this by:

1. Adding an additional `csoAuxilityFnBindings` field to
   `CrucibleSyntaxOverrides`, which stores all of these other functions.
2. Passing these function bindings to `initSimContext` when starting simulation
   in `Ambient.Verifier.SymbolicExecution`.

I've added a test case that demonstrates this fix in action by defining an
override for `strstr` that invokes an auxiliary `compare` function.

Fixes #135.