Support forward declarations in function overrides
This installs the necessary machinery needed to use forward declarations
(see https://github.com/GaloisInc/crucible/pull/997) in syntax overrides.
A high-level description of how this is implemented can be found in
`Note [Resolving forward declarations]` in the new
`A.FunctionOverride.Overrides.ForwardDeclarations` module. As explained in
the Note, the trickiest aspect of all this is pipe-fitting the argument
and result types, so much of this patch revolves around that.

Registering forward declarations in the `test-overrides` command proved
surprisingly tricky, as the code in `runTestOverrides` did not yet have
access to a `FunctionABI`, which is needed to get access to Haskell-defined
overrides. As a result, I had to do a fair bit of incidental refactoring to
construct a `FunctionABI` within `runTestOverrides`, which makes this patch
a bit noisier than I would like in certain places. Oh well.

Fixes #111.