Add {read,write}-c-string overrides
These overrides bridge the gap between C strings and Crucible strings, the
latter of which is more convenient to program with in syntax overrides. Because
the types of these overrides mention `String`, a type which doesn't really have
a direct register equivalent, they can only be invoked from overrides, not
binaries. I've made a note of these in the `README`.

Fixes #138.