Support overrides with more arguments than registers
If a function is supplied more arguments than there are registers, the
remaining arguments will be stored on the stack. When loading arguments before
invoking an override, we now account for this by loading the appropriate number
of arguments from the stack pointer. As described in the new
`Note [Passing arguments to functions]` in `Ambient.FunctionOverride`, this
process is slightly different from ABI to ABI, so we abstract over these ABI
differences in the `functionIntegerArguments` (previously named
`functionIntegerArgumentRegisters`) field of a `FunctionABI`.

This addresses one part of #117 and will lay the foundation for the second
part (supporting varargs).