Mark PLT stubs as trusted function entry points
This is an alternative fix for https://github.com/GaloisInc/macaw/issues/285
(i.e., allowing tail calls into PLT stubs) without needing to rearrange the
order in which `macaw`'s classifiers run (the culprit behind
https://github.com/GaloisInc/macaw/issues/294). Our new approach is to mark
all PLT stub addresses as trusted function entry points so that code discovery
will be more likely to analyze them. See the comments on the new
`lbpTrustedPltEntryPoints` function for the full details.

This bumps the `macaw` submodule to bring in the changes from
https://github.com/GaloisInc/macaw/pull/296, which prompted the need for
changing the behavior on the `ambient-verifier` side in the first place.