loadIntegerStackArguments: Always retrieve memory from GlobalVar
Previously, we captured a `MemImpl` in a closure that we always passed to
`loadIntegerStackArguments`, which led to stale memory reference being loaded.
The fix is simple: always read the memory afresh from a `GlobalVar Mem` each
time. As a consequence, we no longer need to thread around a `mem` throughout
`lookupFunction`, which is nice.

Fixes #148. I will add a test case in a subsequent commit which relies on
this bugfix.