Support varargs in built-in function overrides
Currently, we have a built-in override for the variadic `sprintf` function, and
we would like to add support for `sscanf` soon. Moreover, we have a need to
invoke `sscanf` with differing numbers of arguments. We have danced around this
issue in the past—see #117—but now it is finally time to roll up our sleeves
and handle variadic functions robustly, which this patch accomplishes.

The design is laid out in the new `Note [Varargs]` in
`Ambient.FunctionOverride`. The short version of that Note is that the type of
`functionOverride` now has an additional callback for retrieving variadic
arguments in a similar style to C's `va_arg` macro. This callback is
implemented by some additional machinery in
`Ambient.Override.buildArgumentAssignment`. I have demonstrated that this new
machinery works by updating the built-in `sprintf` override to make use of this
new callback, as well as updating the `sprintf.exe` test cases to show off
examples of invoking the override with varying numbers of arguments. (Support
for `sscanf` will come at a later date.)

Along the way, I needed to bump the `crucible` submodule to bring in the
changes from https://github.com/GaloisInc/crucible/pull/1008, as I needed
access to the internals of `ConversionDirective` in order to load `sprintf`'s
varargs with the appropriate types.

This addresses the second part of #117, finally fixing #117.