Add sscanf override
From the studio that brought you hits such as `sprintf`, presenting: `sscanf`.

`sscanf` is just as hairy as `sprintf`, but instead of putting values into a
string, this time we must extract values out of a string. We do so by:

1. Parsing the format string to determine what conversion directives we have.
   (Luckily, this part can be done entirely using functionality from
   `crucible-llvm` without modifications.)
2. Parsing the input string, extracting out the values corresponding to the
   conversion directives. (See `executeDirectivesScanf`.)
3. Storing each of the values from the previous step into the corresponding
   argument to `sscanf`. (See `scanfOps`.)

This implementation of `sscanf` is far from faithful to the semantics of the
C function of the same name. It is primarily intended to be a quick-and-dirty
way to unblock progress on verifying the R7000 `httpd` example, which makes use
of `sscanf` in control-flow-relevant ways.