Make overrides work for addresses with multiple function names
This generalizes `A.Loader.BinaryConfig.lbpEntryPoints` such that each function
address in the `Map` is now associated with a non-empty list of possible
function names. When checking a function address to see if it has an override,
we now consult all of the possible function names. This allows the verifier to
properly override functions with multiple names (e.g., `open`/`open64` in
`libc`.)

Fixes #142.