Make the first argument to pointer-{read,write} a type, not a Nat
Currently, the type signatures for `pointer-read` and `pointer-write` assume
that the value to read or write is always a `Bitvector`. This precludes the
ability to read or write pointers, however, which is something that arises when
dealing with nested pointers (e.g., `int**`). To make this possible, this patch
generalizes the types of `pointer-{read,write}` a bit. They are now (roughly):

```hs
pointer-read  :: forall (t :: Type) -> Endianness -> Pointer -> t
pointer-write :: forall (t :: Type) -> Endianness -> Pointer -> t -> Unit
```

We say "roughly" above because this doesn't actually work for all possible
`Type`s. The `Type` must be one of the following:

* `Bitvector (8 * w)`, where `w` is a positive number
* One of the aliases for `Bitvector` that we provide
  (e.g., `Int`, `SizeT`, etc.)
* `Pointer`

The possibility of using the `Pointer` type is the key contribution of this
patch. See the `nested-pointers.exe` test case for an example of this in
action. The types above cover all of our current use cases, but if we wish to
add more possible `Type`s in the future, the infrastructure is now in place
for doing so.

This is a breaking change for syntax overrides, and existing uses of
`pointer-{read,write}` will need to be migrated over to the new type. For
instance, suppose you had a syntax override that performs the following to
read from and write to a pointer:

```racket
(let x (pointer-read 4 le ptr))
(pointer-write 4 le ptr x)
```

This can be migrated by picking a suitable type that is 4 bytes in size. For
example, this would work:

```racket
(let x (pointer-read (Bitvector 32) le ptr))
(pointer-write (Bitvector 32) le ptr x)
```

Or, if you are on an architecture where `int` is 4 bytes, this would also work:

```racket
(let x (pointer-read Int le ptr))
(pointer-write Int le ptr x)
```

Fixes #153.