Add get-global-pointer-{addr,named} overrides
These overrides return pointers to a global variable at the specified address
or name, depending on which override you use. This is somewhat novel in that
this is the first time an override needs to look into a binary in order to
function properly, so for this reason, `get-global-pointer-{addr,named}` cannot
be invoked from `test-overrides`. As a result, we now thread through an
additional `FunctionOverrideContext` argument through to the right places in
order to let us know whether an override is being invoked from `verify` (in
which case we have access to the binaries) or `test-overrides` (in which case
we do not).

These overrides are implemented by way of `macaw`'s `doGetGlobal` function.
In order to invoke `doGetGlobal` from an `OverrideSim`, I had to update the
`macaw` submodule to bring in the changes from
https://github.com/GaloisInc/macaw/pull/304, which generalizes the type of
`doGetGlobal` slightly to allow it to be called from an `OverrideSim`.

Another piece of tech that needed to be added to support
`get-global-pointer-named` is the ability to look up _all_ of a binary's
global variable names, not just the dynamic ones. This is now exposed via the
new `lbpGlobalVars` field of `LoadedBinaryPath`. I have renamed the existing
`bcGlobalSymbolMap` field of `BinaryConfig` to `bcDynamicGlobalSymbolMap` to
emphasize the fact that it only covers dynamic global symbols, whereas
`lbpGlobalVars` includes both static _and_ dynamic symbols.