Accommodate changes from https://github.com/GaloisInc/macaw/pull/310
This bumps the `macaw` submodule to bring in the changes from
https://github.com/GaloisInc/macaw/pull/310, which removes some redundant
pointer validity checks from `macaw-symbolic`'s load- and store-related
operations. Because the verifier's test suite counts the number of goals
generated, this has the effect of reducing the overall number of goals in some
test cases. I have confirmed (by manual inspection) that the goals that were
removed with this patch all come from the redundant checks removed in the
`macaw` patch.