Recursively register forward decs during lazy override registration
We were being a bit too lazy when lazily registering overrides. We would
register the forward declarations contained within an override, but not any of
the forward declarations contained within _those_ overrides, resulting in the
error observed in #155. This patch fixes the issue by doing a bit more work and
recursively registering all the forward declarations that an override
transitively depends on.

Fixes #155.