Support R_ARM_COPY relocations
`R_ARM_COPY` is a unique relocation type that references the _value_ of a
global variable defined in another shared library. Contrast this with
`R_ARM_GLOB_DAT` (another relocation type that the verifier supports), which
references the _address_ of another global variable. In order to be able to
simulate this copying behavior in the verifier, we have to add some tweaks in
various places.  These tweaks are described in more detail in `Note [Global
symbols and COPY relocations]` in `Ambient.Loader.ELF.Symbols`, but at a high
level, we must:

1. Add a new `bcSupportedRelocations` field to `BinaryConfig` which tracks
   which relocations are supported, as well as which types of relocations
   (for now, we only distinguish between `COPY` and `GLOB_DAT` relocations),

2. Disambiguate variable definitions using `COPY` relocations from the
   definitions that they reference in
   `Ambient.Loader.ELF.Symbols.addSymbolWithPriority`,

3. Extend `macaw` to support `R_ARM_COPY` relocations, which requires a
   submodule bump to bring in the changes from
   https://github.com/GaloisInc/macaw/pull/311, and

4. Extend `Ambient.Verifier.SymbolicExecution.globalMemoryHooks` to
   be able to handle `COPY` relocations. In order to do so, we must know
   the address of a `COPY` relocation in addition to its name, which
   requires extending the type of `populateGlobal` in `macaw-symbolic` to
   include this address information. This also requires a `macaw` submodule
   bump to bring the necessary changes from
   https://github.com/GaloisInc/macaw/pull/312.

Addresses one bullet point in #98.