Add initial observable events reporting
Partially addresses #123.

This change adds a flag `--log-observable-events <JSON file>` that
writes out any observable events that occur after weird machine entry.
For now it records only function and system calls.

The design is a bit of a hybrid of the WMM and Crucible LLVM annotation
designs.  It's similar to the WMM in that it uses a `SymSequence` to
track the trace of events.  It's similar to Crucible LLVM annotations in
that it assigns an ID to each event and maintains a mapping from IDs to
events.  This is necessary because `SymSequence`s cannot hold arbitrary
datatypes, but they can hold integers.

Next steps from here:
* Add the remaining event types in #123.
* Improve readability of path conditions.
* Stop the trace at weird machine exit.  Currently this reports
  everything from WM entry to the end of simulation.  This should be
  changeable, but as there is no `WeirdMachineExit` `Transition` to
  match `WeirdMachineEntry`, it is going to require some of the
  machinery we need to support the remaining event types that aren't
  `Transition`s.