Relax macaw memop requirements for special pointers
`macaw-symbolic`'s default handling for the `PtrEq` operation treats
comparisons between pointers with different region numbers as undefined
behavior. While this is a suitable choice for some applications, this behavior
proves troublesome for certain types of loops that compare pointer addresses
against raw bitvectors, as we have observed in #159. This slightly relaxes the
requirements that `PtrEq` imposes to allow comparing the stack pointer to other
pointers without requiring that the region numbers be the same. We also relax
the requirements for `PtrToBits`, which also prove troublesome for similar
reasons. For the full explanation, see the new `Note [Special pointer region
numbers]` in `Ambient.Extensions`.

Fixes #159.