Make syscallToFunctionOverride register syscall events
Switching the weird machine executor over to using
`A.V.SymbolicExecution.lookupFunction` reveals an oversight in the way we
register syscall events: whenever we invoke a function wrapper around a
syscall, it is registered as `InvokesFunction`, not `InvokesSyscall`.
Really, we'd like to have both. This performs the necessary refactoring to
make that possible.