Environment variables
This patch adds the necessary machinery to the verifier to allow specifying
environment variables at the command line and allowing their use in both
binaries and overrides. The key additions are:

1. New `--env-var KEY=VALUE` and `--env-var-symbolic KEY[LEN]` command-line
   flags that parse concrete and symbolic environment variable definitions,
   respectively. These are backed by data types in the new `Ambient.EnvVar`
   module.
2. The `A.Verifier.SymbolicExecution.initMainArguments` function now allocates
   the initial value of `envp` (the array of environment variable definitions)
   in addition to `argc` and `argv`. See the `envp.exe` test case.
3. In addition to exposing the environment variables via `envp`, we also expose
   them by way of a global `AMBIENT_environ :: Pointer` variable. This variable
   can be accessed in syntax overrides via `extern` declarations. For example,
   see the `environ.exe` test case, write writes `AMBIENT_environ`'s value to
   `char **environ` at program startup.

One thing that this patch does /not/ accomplish is attempting to simulate the
actual definitions of `getenv`/`setenv`. (Note that the two test cases that I've
added simply define simplistic versions of `getenv` for testing purposes.) That
being said, I suspect that this patch adds most of the pieces needed to make
this possible, so simulating `getenv`/`setenv` is a logical next step.

Fixes #7.