Add free override
For the `default` memory, the implementation is straightforward: just call
`doFree` from `crucible-llvm`. For the `bump-allocator` memory model, I have
made the (possibly contentious) choice of making the `free` override a no-op.
Being any smarter than that would require significantly more work, and given
that `free` was not a major concern for the one exploit where we made use of
the `bump-allocator` memory model, I am inclined not to spend further effort on
it unless someone asks.

I have added two test cases: one for valid uses of `free` (`free-basic.exe`)
and one that tests for some illicit uses of `free` (`free-invalid.exe`). In
order to catch the illicit uses of `free` in the verifier, I needed to tweak
the verifier's elision of undefined behavior to include `free`-specific
behaviors.  See `Note [Undefined behavior and machine code]` for the full
story.

Fixes #183.